+++ vacefron/__init__.py
@@ -1,7 +1,10 @@
+# noinspection PyUnresolvedReferences
 from .classes import Image, RankCard, CrewMateColors
+# noinspection PyUnresolvedReferences
 from .client import Client
+# noinspection PyUnresolvedReferences
 from .errors import BadRequest, HTTPException, InternalServerError, NotFound
 
 __license__ = "MIT"
 __author__ = "Soheab_"
-__version__ = "1.3.0"
+__version__ = "1.4.0"