+++ vacefron/client.py
@@ -1,8 +1,9 @@
 from asyncio import AbstractEventLoop
 from asyncio import get_event_loop
 from random import choice as choose
+from re import search
 from typing import Tuple, Union
-from urllib.parse import quote
+from urllib.parse import quote, urlencode
 
 from aiohttp import ClientSession
 
@@ -25,36 +26,37 @@ async def _get_from_enum(enum_class, value: Union[str, int]):
 
 
 class Client:
-    def __init__(
-            self, session: ClientSession = None, loop: AbstractEventLoop = None
-            ) -> None:
-        self.session = ClientSession(loop = get_event_loop() or loop) or session
-        self._api_url = "https://vacefron.nl/api"
+    __slots__ = ("session", "loop")
 
-    async def _check_url(self, url: str):
-        url = str(url)
-        response = await self.session.get(url = url)
-        if (
-                response.content_type == "application/json"
-                and int((await response.json())["code"]) == 400
-        ):
-            raise BadRequest((await response.json()).get("message"))
+    def __init__(self, *, session: ClientSession = None, loop: AbstractEventLoop = None) -> None:
+        self.session = ClientSession(loop = get_event_loop() or loop) or session
+
+    async def _api_request(self, endpoint: str, params: dict = None):
+        url = f"https://vacefron.nl/api/{endpoint}"
+        if params:
+            encoded_param = urlencode(params, quote_via = quote)
+            url += f"?{encoded_param}"
+        response = await self.session.get(str(url))
+
+        if response.content_type == "application/json" and response.status == 200:
+            return await response.json()
         elif response.status == 200:
-            return url
+            return response
         elif response.status == 400:
-            raise BadRequest((await response.json()).get("message"))
+            raise BadRequest((await response.json()).get("message", "didn't return json..."))
         elif response.status == 404:
-            raise NotFound((await response.json()).get("message"))
+            raise NotFound((await response.json()).get("message", "didn't return json..."))
         elif response.status == 500:
-            raise InternalServerError((await response.json()).get("message"))
+            raise InternalServerError((await response.json()).get("message", "didn't return json..."))
         else:
-            raise HTTPException(response, (await response.json()).get("message"))
+            msg = (await response.json()).get("message", "didn't return json...")
+            raise HTTPException(response, msg)
 
     # Json/URL
 
     async def discord_server(self, creator: bool = False) -> Union[str, Tuple]:
-        api = await self.session.get(self._api_url)
-        support_server = (await api.json()).get("discord_server")
+        json_response = await self._api_request("")
+        support_server = json_response.get("discord_server")
         if creator:
             return support_server, "https://discord.gg/yCzcfju"
 
@@ -62,140 +64,122 @@ class Client:
 
     # Image
 
-    async def distracted_bf(self, boyfriend: str, girlfriend: str, woman: str) -> Image:
-        url = (
-            f"{self._api_url}/distractedbf"
-            f"?boyfriend={str(boyfriend)}"
-            f"&girlfriend={str(girlfriend)}"
-            f"&woman={str(woman)}"
-        )
+    async def batman_slap(self, text: str, text2: str, batman: str = None, robin: str = None) -> Image:
+        params = {"text1": str(text), "text2": str(text2)}
+        if batman:
+            params["batman"] = str(batman)
+        if robin:
+            params["robin"] = str(robin)
 
-        url = await self._check_url(url)
-        return Image(url, self.session)
+        response = await self._api_request("batmanslap", params)
+        return Image(str(response.url), response)
+
+    async def distracted_bf(self, boyfriend: str, girlfriend: str, woman: str) -> Image:
+        response = await self._api_request(
+                "distractedbf", {"boyfriend": str(boyfriend), "girlfriend": str(girlfriend), "woman": str(woman)})
+        return Image(str(response.url), response)
 
     async def car_reverse(self, text: str) -> Image:
-        url = await self._check_url(
-                f"{self._api_url}/carreverse?text={quote(str(text))}"
-                )
-        return Image(url, self.session)
+        response = await self._api_request("carreverse", {"text": str(text)})
+        return Image(str(response.url), response)
 
     async def change_my_mind(self, text: str) -> Image:
-        url = await self._check_url(
-                f"{self._api_url}/changemymind?text={quote(str(text))}"
-                )
-        return Image(url, self.session)
+        response = await self._api_request("changemymind", {"text": str(text)})
+        return Image(str(response.url), response)
 
     async def emergency_meeting(self, text: str) -> Image:
-        url = await self._check_url(
-                f"{self._api_url}/emergencymeeting?text={quote(str(text))}"
-                )
-        return Image(url, self.session)
+        response = await self._api_request("emergencymeeting", {"text": str(text)})
+        return Image(str(response.url), response)
 
-    async def ejected(
-            self,
-            name: str,
-            crewmate: Union[str, int, CrewMateColors] = CrewMateColors.RED,
-            impostor: bool = False,
-            imposter: bool = False,
-            ) -> Image:
+    async def ejected(self, name: str, crewmate: Union[str, int, CrewMateColors] = CrewMateColors.RED,
+                      impostor: bool = False, imposter: bool = False) -> Image:
         impostor = impostor or imposter
         get_color = await _get_from_enum(CrewMateColors, crewmate)
         if get_color is CrewMateColors.RANDOM or not get_color:
             crewmate = choose(list(CrewMateColors)).name
         else:
             crewmate = get_color.name
-        url = await self._check_url(
-                f"{self._api_url}/ejected"
-                f"?name={quote(str(name))}"
-                f"&crewmate={crewmate.lower()}"
-                f"&impostor={impostor}"
-                )
-        return Image(url, self.session)
+
+        response = await self._api_request(
+                "ejected", {"name": str(name), "crewmate": str(crewmate.lower()), "impostor": bool(impostor)})
+        return Image(str(response.url), response)
 
     async def first_time(self, user: str) -> Image:
-        url = await self._check_url(f"{self._api_url}/firsttime?user={str(user)}")
-        return Image(url, self.session)
+        response = await self._api_request("firsttime", {"user": str(user)})
+        return Image(str(response.url), response)
 
     async def grave(self, user: str) -> Image:
-        url = await self._check_url(f"{self._api_url}/grave?user={str(user)}")
-        return Image(url, self.session)
+        response = await self._api_request("grave", {"user": str(user)})
+        return Image(str(response.url), response)
 
     async def iam_speed(self, user: str) -> Image:
-        url = await self._check_url(f"{self._api_url}/iamspeed?user={str(user)}")
-        return Image(url, self.session)
+        response = await self._api_request("iamspeed", {"user": str(user)})
+        return Image(str(response.url), response)
 
     async def i_can_milk_you(self, user: str, user2: str = None) -> Image:
-        url = f"{self._api_url}/icanmilkyou?user1={str(user)}"
+        params = {"user1": str(user)}
         if user2:
-            url += f"&user2={str(user)}"
+            params["user2"] = str(user2)
 
-        url = await self._check_url(url)
-        return Image(url, self.session)
+        response = await self._api_request("icanmilkyou", params)
+        return Image(str(response.url), response)
 
     async def heaven(self, user: str) -> Image:
-        url = await self._check_url(f"{self._api_url}/heaven?user={str(user)}")
-        return Image(url, self.session)
+        response = await self._api_request("heaven", {"user": str(user)})
+        return Image(str(response.url), response)
 
     async def npc(self, text: str, text2: str) -> Image:
-        url = await self._check_url(
-                f"{self._api_url}/npc?text1={quote(str(text))}&text2={quote(str(text2))}"
-                )
-        return Image(url, self.session)
+        response = await self._api_request("npc", {"text1": str(text), "text2": str(text2)})
+        return Image(str(response.url), response)
 
     async def stonks(self, user: str, not_stonks = False) -> Image:
-        url = await self._check_url(f"{self._api_url}/stonks?user={str(user)}")
+        params = {"user": str(user)}
         if not_stonks:
-            url += f"&notstonks=true"
+            params["notstonks"] = True
 
-        return Image(url, self.session)
+        response = await self._api_request("stonks", params)
+        return Image(str(response.url), response)
 
     async def table_flip(self, user: str) -> Image:
-        url = await self._check_url(f"{self._api_url}/tableflip?user={str(user)}")
-        return Image(url, self.session)
+        response = await self._api_request("tableflip", {"user": str(user)})
+        return Image(str(response.url), response)
 
     async def water(self, text: str) -> Image:
-        url = await self._check_url(f"{self._api_url}/water?text={quote(str(text))}")
-        return Image(url, self.session)
+        response = await self._api_request("water", {"text": str(text)})
+        return Image(str(response.url), response)
 
     async def wide(self, image: str) -> Image:
-        url = await self._check_url(f"{self._api_url}/wide?image={str(image)}")
-        return Image(url, self.session)
+        response = await self._api_request("wide", {"image": str(image)})
+        return Image(str(response.url), response)
 
-    async def rank_card(
-            self,
-            username: str,
-            avatar: str,
-            level: int,
-            rank: int,
-            current_xp: int,
-            next_level_xp: int,
-            previous_level_xp: int,
-            custom_background: str = None,
-            xp_color: str = None,
-            is_boosting: bool = False,
-            ) -> RankCard:
+    async def wolverine(self, user: str) -> Image:
+        response = await self._api_request("wolverine", {"user": str(user)})
+        return Image(str(response.url), response)
 
-        custom_background = (
-            str(custom_background).replace("#", "") if custom_background else None
-        )
+    async def rank_card(self, username: str, avatar: str, level: int, rank: int, current_xp: int, next_level_xp: int,
+                        previous_level_xp: int, custom_background: str = None, xp_color: str = None,
+                        is_boosting: bool = False) -> RankCard:
+
+        custom_background = str(custom_background) if custom_background else None
         xp_color = str(xp_color).replace("#", "") if xp_color else None
 
-        card = RankCard(
-                self.session,
-                quote(str(username)),
-                str(avatar),
-                level,
-                rank,
-                current_xp,
-                next_level_xp,
-                previous_level_xp,
-                custom_background,
-                xp_color,
-                is_boosting,
-                )
+        params = {
+            "username": str(username), "avatar": str(avatar), "level": int(level), "rank": int(rank),
+            "currentxp": int(current_xp), "nextlevelxp": int(next_level_xp), "previouslevelxp": int(previous_level_xp)
+            }
+        if custom_background:
+            params['custombg'] = str(custom_background).replace("#", "")
+        if xp_color:
+            if not search(r"^(?:[0-9a-fA-F]{3}){1,2}$", str(xp_color).replace("#", "")):
+                raise BadRequest(
+                        "Invalid HEX value. You're only allowed to enter HEX (0-9 & A-F)"
+                        )
+            params['xpcolor'] = str(xp_color).replace("#", "")
+        if is_boosting:
+            params['isboosting'] = bool(is_boosting)
 
-        await self._check_url(str(card))
-        return card
+        response = await self._api_request("rankcard", params)
+        return RankCard(str(response.url), response, params)
 
     async def close(self) -> None:
         if not self.session.closed: